Rename to LED state and added timeout
