+++ src/main/java/frc/robot/notification/ILedIO.java
@@ -2,7 +2,7 @@ package frc.robot.notification;
 
 public interface ILedIO {
   
-  public void changeColor(LEDStateEnum color);
+  public void changeColor(LEDState color);
 
   public boolean getIsAlive();
